+++ feature/drawing/src/main/java/com/dev/philo/fillsketch/feature/drawing/viewmodel/DrawingViewModel.kt
@@ -63,7 +63,7 @@ class DrawingViewModel @Inject constructor(
 
     fun fetchDrawingUiState(
         sketchType: Int,
-        drawingResultId: Int,
+        drawingResultId: Long,
         recommendImageBitmap: ImageBitmap,
         outlineImageBitmap: ImageBitmap,
     ) {
