+++ feature/home/src/main/java/com/dev/philo/fillsketch/feature/home/model/SketchListUiEvent.kt
@@ -1,5 +1,5 @@
 package com.dev.philo.fillsketch.feature.home.model
 
 sealed interface SketchListUiEvent {
-    data class NavigateToDrawing(val sketchType: Int, val drawingResultId: Int) : SketchListUiEvent
+    data class NavigateToDrawing(val sketchType: Int, val drawingResultId: Long) : SketchListUiEvent
 }
